Centralise access key logic
