+++ hq/app/schedule/IamFlaggedUsers.scala
@@ -1,7 +1,6 @@
 package schedule
 
-import config.Config.{iamHumanUserRotationCadence, iamMachineUserRotationCadence}
-import logic.DateUtils
+import logic.VulnerableAccessKeys
 import model._
 import play.api.Logging
 import schedule.IamDeadline.filterUsersToAlert
@@ -45,13 +44,10 @@ object IamFlaggedUsers extends Logging {
   }
 
   def findOldAccessKeys(credsReport: CredentialReportDisplay): CredentialReportDisplay = {
-    val filteredMachines = credsReport.machineUsers.filter(user => hasOutdatedMachineKey(List(user.key1, user.key2)))
-    val filteredHumans = credsReport.humanUsers.filter(user => hasOutdatedHumanKey(List(user.key1, user.key2)))
+    val filteredMachines = credsReport.machineUsers.filter(user => VulnerableAccessKeys.hasOutdatedMachineKey(List(user.key1, user.key2)))
+    val filteredHumans = credsReport.humanUsers.filter(user => VulnerableAccessKeys.hasOutdatedHumanKey(List(user.key1, user.key2)))
     credsReport.copy(machineUsers = filteredMachines, humanUsers = filteredHumans)
   }
-
-  def hasOutdatedHumanKey(keys: List[AccessKey]): Boolean = keys.exists(key => DateUtils.dayDiff(key.lastRotated).getOrElse(1L) > iamHumanUserRotationCadence)
-  def hasOutdatedMachineKey(keys: List[AccessKey]): Boolean = keys.exists(key => DateUtils.dayDiff(key.lastRotated).getOrElse(1L) > iamMachineUserRotationCadence)
 
   def findMissingMfa(credsReport: CredentialReportDisplay): CredentialReportDisplay = {
     val removeMachineUsers = credsReport.machineUsers.filterNot(_.username == "")
